add category product insert

+++ model/model.go
@@ -29,7 +29,7 @@ type Image struct {
 }
 
 type CategoryProduct struct {
-	ID         uint      `gorm:"primary_key;column:id" json:"id"`
+	ID         uint64    `gorm:"primary_key;column:id" json:"id"`
 	ProductID  uint      `gorm:"column:product_id" json:"product_id"`
 	CategoryID uint      `gorm:"column:category_id" json:"category_id"`
 	CreatedAt  time.Time `gorm:"default:CURRENT_TIMESTAMP" json:"created_at"`
